[ remove ]Cocos2d-x Physics仕様に作った衝突判定関係のクラスを削除しました。

+++ Project/Classes/GameMain/GamePlay/Object/Wall/WallManager.h
@@ -3,7 +3,7 @@
 
 #include "cocos2d.h"
 
-enum class ContactCategory : int;
+enum class ContactCategory : unsigned short;
 
 /*------------------------------------------------------------*/
 //	@class		：	WallManager
